apps/eks-to-azure/requirements.txt
azure-identity==1.15.0
azure-mgmt-resource==23.0.1
boto3==1.34.34
PyJWT==2.8.0